backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.65.0
supabase==2.28.0
python-dotenv==1.1.0
requests==2.32.3
